use calculator display hook implemented
